MTSDK-32: New automation for typing indicators
* New test to verify that we can send a typing indicator and that we can receive one.
* Moving the TestContentController to a separate file.
* New API request for sending a typing indicator from the agent that can be received in tests.
